added torch DataSet

+++ moge/generator/sequences.py
@@ -171,6 +171,6 @@ class MultiSequenceTokenizer(SequenceTokenizer):
         encoded = pad_sequences(encoded, maxlen=maxlen, padding=self.padding,
                                 truncating=np.random.choice(
                                     ["post", "pre"]) if self.truncating == "random" else self.truncating,
-                                dtype="int8")
+                                dtype="long")
 
         return encoded
